fonts.conf: move and update
Add minimal fonts.conf configuration that uses the auto-hinter rather than the
BCI.  For some reason, I cannot get the BCI to render the fonts as smoothly as
the auto-hinter, even though the BCI is supposed to be a better hinter.

fontconfig 1.10.1 deprecated the old path of ${HOME}/.fonts.conf in favour of
${XDG_CONFIG_DIR}/fontconfig/fonts.conf.  Move the configuration in light of
such.