vimrc: hoist clang-format support into vimrc
This generalises and hoists a long time per-host configuration bit into the
vimrc.  Although the per-host setting is still required for the path to
clang-format, if one is in the path, that is sufficient for this magic to work.
This is in preparation for swift-format support to be merged into this codepath.
We now detect the filetype and if the format is one which clang-format supports,
map <leader>f to format code.

This will also allow us to bind future helpers (e.g. <leader>r for variable
renaming).