tmux.conf: adjust colors, style, settings
Change the colourscheme in tmux to make it blend with the background
color in vim with spacegrey.  Make the colours more monochrome to make
it fade out of view.

Make tmux renumber the windows when one is closed.  Force tmux to resize
the window to the local view.  This may make it slightly more annoying
for shared sessions, but should be helpful normally.